Move BukkitMovingObject
