+++ cannons-bukkit/src/main/java/at/pavlov/cannons/Aiming.java
@@ -10,7 +10,7 @@ import at.pavlov.cannons.cannon.CannonManager;
 import at.pavlov.cannons.cannon.DesignStorage;
 import at.pavlov.cannons.config.Config;
 import at.pavlov.cannons.config.UserMessages;
-import at.pavlov.cannons.container.BukkitMovingObject;
+import at.pavlov.bukkit.container.BukkitMovingObject;
 import at.pavlov.cannons.container.Target;
 import at.pavlov.cannons.dao.AsyncTaskManager;
 import at.pavlov.cannons.event.CannonLinkAimingEvent;
